requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-apache-kafka
confluent-kafka==2.0.2
requests
kafka-python==2.0.2